fix(web): isolate Edge Runtime middleware from Node-only auth + test helpers
PR #34's verify CI was failing with `UnhandledSchemeError: Reading from
"node:crypto" is not handled by plugins` and pglite dynamic-code-evaluation
errors on the Next.js build. Root cause: `middleware.ts` (Edge Runtime)
transitively imported the full Stage 1 composition root via
`./src/server/auth/index.ts → stage1-runtime.ts → @as-comms/domain + db/test-helpers`,
which pulls `node:crypto` (domain/notes.ts) and PGlite into the Edge bundle.

Fix: split auth + composition root into Edge-safe and Node-only halves,
following the standard Auth.js v5 Edge-compatibility pattern.

- `apps/web/src/server/auth/config.ts` (new): Edge-safe Auth.js config —
  providers, pages, trustHost. No adapter, no DB-dependent callbacks.
- `apps/web/src/server/auth/index.ts`: now imports the edge config and
  extends it with DrizzleAdapter + database-session strategy + role-aware
  session / signIn callbacks for Node-runtime consumers.
- `apps/web/middleware.ts`: creates a local NextAuth instance from
  `authEdgeConfig` only; no transitive Node imports.
- `apps/web/src/server/stage1-runtime.test-support.ts` (new): holds
  `createStage1WebTestRuntime`, `Stage1WebTestRuntime`, and the
  `TestStage1Context` re-export — test-only APIs that previously
  lived in `stage1-runtime.ts` and re-exported PGlite via
  `@as-comms/db/test-helpers`.
- `apps/web/src/server/stage1-runtime.ts`: test-helper re-export and
  `createStage1WebTestRuntime` function removed. Production runtime stays
  importable from Edge Runtime code paths without pulling PGlite.
- `apps/web/tests/unit/inbox-stage1-helpers.ts`: updated import to the
  new `.test-support` module.
- `scripts/boundary-check.mjs`: composition-root allowlist extended to
  include `stage1-runtime.test-support.ts` so test-only code can still
  import `@as-comms/db/test-helpers`.

Verified locally with `pnpm boundaries`, `pnpm --filter @as-comms/web
typecheck`, and `pnpm --filter @as-comms/web build` (middleware compiles
to a 87.2 kB Edge chunk).

Middleware-vs-full session note: the middleware instance uses Auth.js's
default JWT-cookie decoding (no adapter in Edge) and only gates on cookie
presence + dev bypass. Route handlers and Server Actions still use the
full `auth()` from `./index.ts` with the DrizzleAdapter + database
session strategy for authoritative role checks. The settings-bundle's
"server-owned sessions" rule still holds — database sessions remain the
authoritative store.
